+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Fix mods not behaving between MC version changes
 - Fix sort box on Add Mods dialog not showing
 - Fix updating Curse mod not selecting current version by default
+- Remove Minecraft status checking as api no longer returns correct